refactor: updated the react-hook form and zustand for better reusability
